Generalize `shallowlyEqual`
